+++ Dyna/docs/release-notes.md
@@ -4,6 +4,7 @@
 
 ## Upcoming release
 
+* The Bastion status now gives as much shield as possible, before giving temp shield.
 * Reversed the Burst Charge card upgrades.
 
 ## 1.0.0